refactor(cli-runtime): split session specs by behavior
